--- bbyy_jet_classifier/plotting/__init__.py
@@ -1,1 +1,0 @@
-from plot import *
