+++ run_classifier.py
@@ -2,7 +2,8 @@
 import argparse
 import logging
 import os
-from bbyy_jet_classifier import strategies, plotting, process_data
+from bbyy_jet_classifier import strategies, process_data
+from bbyy_jet_classifier.plotting import plotinputs, plotoutputs, roc
 
 if __name__ == "__main__":
     # -- Configure logging
@@ -41,14 +42,14 @@ if __name__ == "__main__":
         logging.getLogger("RunClassifier").info("Preparing to train with {}% of events and then test with the remainder".format(int(100 * args.ftrain)))
 
         #-- Plot training distributions
-        plotting.input_distributions(ML_strategy, classification_variables, X_train, y_train, w_train, process="training")  # plot the feature distributions
+        plotinputs.input_distributions(ML_strategy, classification_variables, X_train, y_train, w_train, process="training")  # plot the feature distributions
 
         # -- Train classifier
         ML_strategy.train(X_train, y_train, w_train, classification_variables, variable_dict)
 
         # -- Plot the classifier output as tested on the training set (only useful if you care to check the performance on the training set)
         yhat_train = ML_strategy.test(X_train, y_train, w_train, classification_variables, process="training")
-        plotting.classifier_output(ML_strategy, yhat_train, y_train, w_train, process="training", fileID=args.input.replace(".root", "").split("/")[-1])
+        plotoutputs.classifier_output(ML_strategy, yhat_train, y_train, w_train, process="training", fileID=args.input.replace(".root", "").split("/")[-1])
 
     else:
         logging.getLogger("RunClassifier").info("Preparing to use 100% of sample as testing input")
@@ -56,19 +57,19 @@ if __name__ == "__main__":
     # -- Testing!
     if args.ftrain < 1:
         #-- Plot input testing distributions
-        plotting.input_distributions(ML_strategy, classification_variables, X_test, y_test, w_test, process="testing")
+        plotinputs.input_distributions(ML_strategy, classification_variables, X_test, y_test, w_test, process="testing")
         
         # -- TEST
         yhat_test = ML_strategy.test(X_test, y_test, w_test, classification_variables, process="testing")
 
         # -- Plot output testing distributions from classifier and old strategies
-        plotting.classifier_output(ML_strategy, yhat_test, y_test, w_test, process="testing", fileID=args.input.replace(".root", "").split("/")[-1])
-        plotting.old_strategy(ML_strategy.output_directory, mHmatch_test, y_test, w_test, "mHmatch")
-        plotting.old_strategy(ML_strategy.output_directory, pThigh_test, y_test, w_test, "pThigh")
+        plotoutputs.classifier_output(ML_strategy, yhat_test, y_test, w_test, process="testing", fileID=args.input.replace(".root", "").split("/")[-1])
+        plotoutputs.old_strategy(ML_strategy.output_directory, mHmatch_test, y_test, w_test, "mHmatch")
+        plotoutputs.old_strategy(ML_strategy.output_directory, pThigh_test, y_test, w_test, "pThigh")
 
         # -- Visualize performance by displaying the ROC curve from the selected ML strategy and comparing it with the old strategies
         logging.getLogger("RunClassifier").info("Plotting ROC curves...")
-        plotting.roc(ML_strategy, mHmatch_test, pThigh_test, yhat_test, y_test, w_test)
+        roc.signal_eff_bkg_rejection(ML_strategy, mHmatch_test, pThigh_test, yhat_test, y_test, w_test)
 
     else:
         logging.getLogger("RunClassifier").info("100% of the sample was used for training -- no independent testing can be performed.")
